requirements.txt
beautifulsoup4==4.9.3
python-telegram-bot==11.1.0
requests==2.25.1